moved viewType to string resource converta
